Mini refactor, need to fix mocks_temp.go
