+++ cmd/frontend/internal/httpapi/api_test.go
@@ -30,7 +30,6 @@ func newTest() *httptestutil.Client {
 			GitLabWebhook:             enterpriseServices.GitLabWebhook,
 			BitbucketServerWebhook:    enterpriseServices.BitbucketServerWebhook,
 			BitbucketCloudWebhook:     enterpriseServices.BitbucketCloudWebhook,
-			SyncGitHubWebhook:         enterpriseServices.SyncGitHubWebhook,
 			NewCodeIntelUploadHandler: enterpriseServices.NewCodeIntelUploadHandler,
 			NewComputeStreamHandler:   enterpriseServices.NewComputeStreamHandler,
 		},
